完成初始数据持久化，数据更新待完善

+++ siever-cli/src/main/java/com/huanhe_tech/cli/queue/FilteredByTypeFlowingSymbol.java
@@ -1,14 +1,14 @@
 package com.huanhe_tech.cli.queue;
 
 final public class FilteredByTypeFlowingSymbol extends FlowingSymbolObj{
-    private final int conid;
+    private final long conid;
 
-    public FilteredByTypeFlowingSymbol(int id, String symbol, int conid) {
+    public FilteredByTypeFlowingSymbol(int id, String symbol, long conid) {
         super(id, symbol);
         this.conid = conid;
     }
 
-    public int getConid() {
+    public long getConid() {
         return conid;
     }
 }
